requirements.txt
Flask==3.0.0
PyMuPDF==1.23.12
Pillow==10.1.0
Werkzeug==3.0.1
numpy==1.24.3
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
requests==2.31.0 